feat : edit delete auth

+++ routes/campground.js
@@ -67,6 +67,14 @@ router.post('/', isLogedin ,validatecg ,catchAsync( async (req,res) => {
 
 router.put('/:id', catchAsync( async(req,res,next) => {
     const {id} = req.params;
+    
+    const aa = await Campground.findById(id);
+    
+    if(!aa.author.equals(req.user._id)){
+        req.flash('error','You don\'t have permission to edit' )
+        return  res.redirect(`/cg/${id}`);
+    }
+
     const a = await Campground.findByIdAndUpdate(id,req.body);
     req.flash('success','succesfully updated mapper');
     res.redirect(`/cg/${id}`);
@@ -75,12 +83,28 @@ router.put('/:id', catchAsync( async(req,res,next) => {
 
 router.get('/:id/edit',isLogedin ,catchAsync( async (req,res) => {
     const {id} = req.params;
+    
+    const aa = await Campground.findById(id);
+    
+    if(!aa.author.equals(req.user._id)){
+        req.flash('error','You don\'t have permission to edit' )
+        return  res.redirect(`/cg/${id}`);
+    }
+
     const a = await Campground.findById(id);
     res.render('cg/edit',{a});
 }));
 
 router.delete('/:id', catchAsync( async (req,res) => {
     const {id} = req.params;
+    
+    const aa = await Campground.findById(id);
+    
+    if(!aa.author.equals(req.user._id)){
+        req.flash('error','You don\'t have permission to edit' )
+        return  res.redirect(`/cg/${id}`);
+    }
+
     await Campground.findByIdAndDelete(id);
     req.flash('success','succesfully deleted mapper');
     res.redirect(`/cg`);
